requirement.txt
decorator==4.3.0
Django==2.0.6
django-filter==1.1.0
djangorestframework==3.8.2




